error handling

+++ objective.py
@@ -27,7 +27,7 @@ def contrastive_loss(y_train, x_train):
   tf.compat.v1.enable_eager_execution()
   left = tf.keras.Input( shape=(32, 32, 3), dtype=tf.dtypes.float32)
   right = tf.keras.Input( shape=(32, 32, 3), dtype=tf.dtypes.float32)
-  label = tf.keras.Input( shape=(1), dtype=tf.dtypes.int32)
+  label = tf.keras.Input( shape=(1), dtype=tf.dtypes.float32)
 
   margin = 0.2
   print(tf.shape(left))
